nova-meta-harness/pyproject.toml
[project]
name = "nova-meta-harness-nova"
version = "0.1.0"
description = "Nova integration package for the Meta-Harness framework"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "jinja2>=3.0",
    "diff-match-patch>=2024",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]
claude = ["anthropic>=0.28"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
